refactor replace categories contextAPI with categories redux reducer
