+++ src/components/ProductCard/ProductCard.tsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import styled from 'styled-components';
 import CartContext from '../../contexts/cart';
-import { ProductType } from '../../contexts/categories';
 import Button from '../Button/Button';
 import { COLORS } from '../../constants';
+import { ProductType } from '../../types/product';
 
 const ProductCard = ({ product }: { product: ProductType }) => {
   const { name, imageSrc, price } = product;
