+++ src/store/cart.tsx
@@ -1,4 +1,4 @@
-import { ProductType } from '../contexts/categories';
+import { ProductType } from '../types/product';
 
 export const CART_ACTION_TYPES = {
   SET_CART_ITEMS: 'SET_CART_ITEMS',
