+++ src/types/category.ts
@@ -1,0 +1,9 @@
+import { ProductType } from './product';
+
+export type CategoryType = {
+  id: number;
+  title: string;
+  path: string;
+  imageSrc: string;
+  items: ProductType[];
+};
