+++ src/types/product.ts
@@ -1,0 +1,7 @@
+export type ProductType = {
+  id: number;
+  name: string;
+  imageSrc: string;
+  price: number;
+  quantity?: number;
+};
