+++ src/types/store.ts
@@ -1,0 +1,11 @@
+import { UserType } from '../types/user';
+import { CategoryType } from '../types/category';
+
+export type StoreType = {
+  user: {
+    currentUser: UserType;
+  };
+  categories: {
+    categories: CategoryType[];
+  };
+};
