+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export type UserType = {
+  uid: string;
+  displayName: string;
+  email: string;
+};
